+++ packages/demo-gatsby/gatsby-config.js
@@ -21,12 +21,12 @@ const REPO_ABSOLUTE_PATH = path.join(process.cwd(), "../..")
 
 module.exports = {
   siteMetadata: {
-    title: `Gatsby Starter Blog`,
-    author: `Kyle Mathews`,
-    description: `A starter blog demonstrating what Gatsby can do.`,
+    title: `TinaCMS Starter Blog`,
+    author: `Nolan Philipps`,
+    description: `A starter blog demonstrating what TinaCMS can do.`,
     siteUrl: `https://gatsby-starter-blog-demo.netlify.com/`,
     social: {
-      twitter: `kylemathews`,
+      twitter: `ncphi`,
     },
   },
   plugins: [
@@ -39,6 +39,9 @@ module.exports = {
         sidebar: {
           hidden: process.env.NODE_ENV === "production",
           position: "displace",
+        },
+        toolbar: {
+          hidden: false,
         },
         plugins: [
           "gatsby-tinacms-teams",