+++ packages/@tinacms/react-modals/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.2-alpha.0](https://github.com/tinacms/tinacms/compare/@tinacms/react-modals@0.1.1...@tinacms/react-modals@0.1.2-alpha.0) (2020-04-06)
+
+**Note:** Version bump only for package @tinacms/react-modals
+
+
+
+
+
+## [0.1.1](https://github.com/tinacms/tinacms/compare/@tinacms/react-modals@0.1.0...@tinacms/react-modals@0.1.1) (2020-04-06)
+
+**Note:** Version bump only for package @tinacms/react-modals
+
+
+
+
+
 # 0.1.0 (2020-03-23)