+++ packages/@tinacms/react-sidebar/CHANGELOG.md
@@ -2,3 +2,34 @@
 
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+## [0.1.2-alpha.0](https://github.com/tinacms/tinacms/compare/@tinacms/react-sidebar@0.1.1...@tinacms/react-sidebar@0.1.2-alpha.0) (2020-04-06)
+
+**Note:** Version bump only for package @tinacms/react-sidebar
+
+
+
+
+
+## [0.1.1](https://github.com/tinacms/tinacms/compare/@tinacms/react-sidebar@0.1.0...@tinacms/react-sidebar@0.1.1) (2020-04-06)
+
+
+### Bug Fixes
+
+* styles dep and make react-based deps peers ([62bf2c1](https://github.com/tinacms/tinacms/commit/62bf2c1))
+
+
+
+
+
+# 0.1.0 (2020-04-06)
+
+
+### Bug Fixes
+
+* don't use sitewrapper when sidebar is hidden ([dc8a91d](https://github.com/tinacms/tinacms/commit/dc8a91d))
+
+
+### Features
+
+* introduce @tinacms/react-sidebar ([4281d36](https://github.com/tinacms/tinacms/commit/4281d36))