+++ packages/gatsby-plugin-tinacms/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.34-alpha.0](https://github.com/tinacms/tinacms/compare/gatsby-plugin-tinacms@0.1.33...gatsby-plugin-tinacms@0.1.34-alpha.0) (2020-04-06)
+
+**Note:** Version bump only for package gatsby-plugin-tinacms
+
+
+
+
+
+## [0.1.33](https://github.com/tinacms/tinacms/compare/gatsby-plugin-tinacms@0.1.32...gatsby-plugin-tinacms@0.1.33) (2020-04-06)
+
+**Note:** Version bump only for package gatsby-plugin-tinacms
+
+
+
+
+
+## [0.1.32](https://github.com/tinacms/tinacms/compare/gatsby-plugin-tinacms@0.1.31...gatsby-plugin-tinacms@0.1.32) (2020-04-06)
+
+**Note:** Version bump only for package gatsby-plugin-tinacms
+
+
+
+
+
 ## [0.1.31](https://github.com/tinacms/tinacms/compare/gatsby-plugin-tinacms@0.1.30...gatsby-plugin-tinacms@0.1.31) (2020-03-30)
 
 **Note:** Version bump only for package gatsby-plugin-tinacms