+++ packages/gatsby-tinacms-mdx/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.10-alpha.0](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-mdx@0.1.9...gatsby-tinacms-mdx@0.1.10-alpha.0) (2020-04-06)
+
+**Note:** Version bump only for package gatsby-tinacms-mdx
+
+
+
+
+
+## [0.1.9](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-mdx@0.1.8...gatsby-tinacms-mdx@0.1.9) (2020-04-06)
+
+**Note:** Version bump only for package gatsby-tinacms-mdx
+
+
+
+
+
+## [0.1.8](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-mdx@0.1.7...gatsby-tinacms-mdx@0.1.8) (2020-04-06)
+
+**Note:** Version bump only for package gatsby-tinacms-mdx
+
+
+
+
+
 ## [0.1.7](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-mdx@0.1.6...gatsby-tinacms-mdx@0.1.7) (2020-03-30)
 
 **Note:** Version bump only for package gatsby-tinacms-mdx