+++ packages/react-tinacms-inline/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.5-alpha.0](https://github.com/tinacms/tinacms/compare/react-tinacms-inline@0.4.4...react-tinacms-inline@0.4.5-alpha.0) (2020-04-06)
+
+**Note:** Version bump only for package react-tinacms-inline
+
+
+
+
+
+## [0.4.4](https://github.com/tinacms/tinacms/compare/react-tinacms-inline@0.4.3...react-tinacms-inline@0.4.4) (2020-04-06)
+
+**Note:** Version bump only for package react-tinacms-inline
+
+
+
+
+
+## [0.4.3](https://github.com/tinacms/tinacms/compare/react-tinacms-inline@0.4.2...react-tinacms-inline@0.4.3) (2020-04-06)
+
+**Note:** Version bump only for package react-tinacms-inline
+
+
+
+
+
 ## [0.4.2](https://github.com/tinacms/tinacms/compare/react-tinacms-inline@0.4.1...react-tinacms-inline@0.4.2) (2020-03-30)
 
 **Note:** Version bump only for package react-tinacms-inline